Comprehensive refactor of avant scripts, resyncing SCSS with CSS.
