+++ scripts/avant.js
@@ -1,2298 +1,259 @@
-// Avant - Game System for FoundryVTT v12/v13 Compatible
+/**
+ * @fileoverview Avant Native System - Main System File
+ * @version 2.0.0
+ * @author Avant Development Team
+ * @description Main entry point for the Avant Native FoundryVTT system with v12/v13 compatibility
+ */
 
-console.log("Avant | Loading game system...");
+// Core imports - Component-based architecture
+import { CompatibilityUtils } from './utils/compatibility.js';
+import { ValidationUtils } from './utils/validation.js';
 
-// Import theme manager
+// Data models
+import { AvantActorData } from './data/actor-data.js';
+import { AvantActionData, AvantFeatureData, AvantTalentData, AvantAugmentData, AvantWeaponData, AvantArmorData, AvantGearData } from './data/item-data.js';
+
+// Sheet classes
+import { AvantActorSheet } from './sheets/actor-sheet.js';
+import { AvantItemSheet } from './sheets/item-sheet.js';
+
+// Dialog classes
+import { AvantRerollDialog } from './dialogs/reroll-dialog.js';
+
+// Chat functionality
+import { AvantChatContextMenu } from './chat/context-menu.js';
+
+// Theme manager
 import { AvantThemeManager } from './themes/theme-manager.js';
 
+console.log("Avant | Loading Avant Native System...");
+CompatibilityUtils.log("System initialization started");
+
 /**
- * Actor Data Model for Avant - v12/v13 Compatible
- * @extends {foundry.abstract.DataModel}
+ * Initialize the Avant Native system
+ * This is the main entry point for system setup
  */
-class AvantActorData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            // Basic Character Info
-            level: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 1,
-                min: 1
-            }),
-            fortunePoints: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 3,
-                min: 0
-            }),
-            ancestry: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            lineage: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            culture: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            vocation: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            background: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            languages: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            
-            // Abilities (Core Stats)
-            abilities: new fields.SchemaField({
-                might: new fields.SchemaField({
-                    value: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 10,
-                        min: 1
-                    }),
-                    mod: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 0
-                    })
-                }),
-                grace: new fields.SchemaField({
-                    value: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 10,
-                        min: 1
-                    }),
-                    mod: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 0
-                    })
-                }),
-                intellect: new fields.SchemaField({
-                    value: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 10,
-                        min: 1
-                    }),
-                    mod: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 0
-                    })
-                }),
-                focus: new fields.SchemaField({
-                    value: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 10,
-                        min: 1
-                    }),
-                    mod: new fields.NumberField({
-                        required: true,
-                        nullable: false,
-                        integer: true,
-                        initial: 0
-                    })
-                })
-            }),
-            
-            // Skills System (12 skills from prototype)
-            skills: new fields.SchemaField({
-                debate: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                discern: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                endure: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                finesse: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                force: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                command: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                charm: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                hide: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                inspect: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                intuit: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                recall: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                }),
-                surge: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0
-                })
-            }),
-            
-            // Power Points System
-            powerPoints: new fields.SchemaField({
-                value: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                max: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                limit: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            }),
-            
-            // Expertise Points (for Gear tab)
-            expertisePoints: new fields.SchemaField({
-                total: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                spent: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                remaining: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            }),
-            
-            // Combat Stats
-            health: new fields.SchemaField({
-                value: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 25,
-                    min: 0
-                }),
-                max: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 25,
-                    min: 0
-                }),
-                temp: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            }),
-            
-            defenseThreshold: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 11,
-                min: 1
-            }),
-            
-            // Character Details
-            biography: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            notes: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            })
+Hooks.once('init', async function() {
+    console.log('Avant | Initializing Avant Native system');
+    
+    try {
+        // Initialize theme manager - register settings first
+        AvantThemeManager.registerSettings();
+        CompatibilityUtils.log("Theme manager settings registered");
+        
+        // Configure data models for the system
+        const actors = CompatibilityUtils.getActorsCollection();
+        const items = CompatibilityUtils.getItemsCollection();
+        
+        // Register Actor data models
+        CONFIG.Actor.documentClass = AvantActor;
+        CONFIG.Actor.dataModels = {
+            character: AvantActorData,
+            npc: AvantActorData,
+            vehicle: AvantActorData
         };
+        
+        // Register Item data models
+        CONFIG.Item.documentClass = AvantItem;
+        CONFIG.Item.dataModels = {
+            action: AvantActionData,
+            feature: AvantFeatureData,
+            talent: AvantTalentData,
+            augment: AvantAugmentData,
+            weapon: AvantWeaponData,
+            armor: AvantArmorData,
+            gear: AvantGearData
+        };
+        
+        // Register sheet application classes
+        actors.unregisterSheet("core", CompatibilityUtils.getActorSheetClass());
+        actors.registerSheet("avant", AvantActorSheet, { makeDefault: true });
+        
+        items.unregisterSheet("core", CompatibilityUtils.getItemSheetClass());
+        items.registerSheet("avant", AvantItemSheet, { makeDefault: true });
+        
+        CompatibilityUtils.log("Data models and sheets registered");
+        
+        // Preload Handlebars templates
+        const loadTemplates = CompatibilityUtils.getLoadTemplatesFunction();
+        await loadTemplates([
+            "systems/avant/templates/actor-sheet.html",
+            "systems/avant/templates/item-sheet.html",
+            "systems/avant/templates/reroll-dialog.html",
+            "systems/avant/templates/item/item-action-sheet.html",
+            "systems/avant/templates/item/item-feature-sheet.html",
+            "systems/avant/templates/item/item-talent-sheet.html",
+            "systems/avant/templates/item/item-augment-sheet.html",
+            "systems/avant/templates/item/item-weapon-sheet.html",
+            "systems/avant/templates/item/item-armor-sheet.html",
+            "systems/avant/templates/item/item-gear-sheet.html"
+        ]);
+        
+        CompatibilityUtils.log("Templates preloaded");
+        
+        console.log('Avant | System initialization complete');
+        
+    } catch (error) {
+        console.error('Avant | Error during system initialization:', error);
+        CompatibilityUtils.log(`System initialization failed: ${error.message}`, 'error');
+    }
+});
+
+/**
+ * Once the entire VTT framework is initialized, set up additional features
+ */
+Hooks.once('ready', async function() {
+    console.log('Avant | System ready - setting up additional features');
+    
+    try {
+        // Initialize chat context menu for Fortune Point rerolls
+        AvantChatContextMenu.addContextMenuListeners();
+        CompatibilityUtils.log("Chat context menu initialized");
+        
+        // Create and initialize theme manager instance
+        if (!game.avant) game.avant = {};
+        game.avant.themeManager = new AvantThemeManager();
+        await game.avant.themeManager.init();
+        CompatibilityUtils.log("Theme manager instance created and initialized");
+        
+        console.log('Avant | System ready and fully configured');
+        
+    } catch (error) {
+        console.error('Avant | Error during ready setup:', error);
+        CompatibilityUtils.log(`Ready setup failed: ${error.message}`, 'error');
+    }
+});
+
+/**
+ * Extended Actor class for Avant Native
+ * @extends {Actor}
+ */
+class AvantActor extends Actor {
+    /**
+     * Prepare actor data
+     * Calculate derived values and ensure data consistency
+     * @override
+     */
+    prepareData() {
+        super.prepareData();
+        
+        // Validate actor data
+        this.system = ValidationUtils.validateActorData(this.system);
     }
     
     /**
-     * Get skill ability mapping
-     * @returns {Object} Mapping of skills to abilities
+     * Prepare derived data
+     * Called after base data preparation
+     * @override
      */
-    static getSkillAbilities() {
-        return {
-            debate: 'intellect',
-            discern: 'focus',
-            endure: 'focus',
-            finesse: 'grace',
-            force: 'might',
-            command: 'might',
-            charm: 'grace',
-            hide: 'grace',
-            inspect: 'intellect',
-            intuit: 'focus',
-            recall: 'intellect',
-            surge: 'might'
-        };
-    }
-}
-
-/**
- * Item Data Models - v12/v13 Compatible
- */
-
-/**
- * Action Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantActionData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            ability: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "might",
-                choices: ["might", "grace", "intellect", "focus"]
-            }),
-            difficulty: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 11,
-                min: 1
-            })
-        };
-    }
-}
-
-/**
- * Feature Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantFeatureData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            source: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            category: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "general"
-            }),
-            uses: new fields.SchemaField({
-                value: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                max: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            })
-        };
-    }
-}
-
-/**
- * Talent Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantTalentData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            powerPointCost: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 0,
-                min: 0
-            }),
-            prerequisites: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            uses: new fields.SchemaField({
-                value: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                max: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            })
-        };
-    }
-}
-
-/**
- * Augment Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantAugmentData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            powerPointCost: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 0,
-                min: 0
-            }),
-            augmentType: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "enhancement"
-            }),
-            prerequisites: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            uses: new fields.SchemaField({
-                value: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                }),
-                max: new fields.NumberField({
-                    required: true,
-                    nullable: false,
-                    integer: true,
-                    initial: 0,
-                    min: 0
-                })
-            })
-        };
-    }
-}
-
-/**
- * Weapon Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantWeaponData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            ability: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "might",
-                choices: ["might", "grace", "intellect", "focus"]
-            }),
-            modifier: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 0
-            }),
-            damageDie: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "1d6"
-            }),
-            damageType: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            range: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "Melee"
-            }),
-            properties: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            weight: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 1,
-                min: 0
-            }),
-            cost: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 0,
-                min: 0
-            }),
-            threshold: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 11,
-                min: 1
-            })
-        };
-    }
-}
-
-/**
- * Armor Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantArmorData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            ability: new fields.StringField({
-                required: true,
-                blank: false,
-                initial: "grace",
-                choices: ["might", "grace", "intellect", "focus"]
-            }),
-            modifier: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 0
-            }),
-            damageReduction: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 0,
-                min: 0
-            }),
-            properties: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            weight: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 1,
-                min: 0
-            }),
-            cost: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 0,
-                min: 0
-            }),
-            threshold: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 11,
-                min: 1
-            })
-        };
-    }
-}
-
-/**
- * Gear Item Data Model
- * @extends {foundry.abstract.DataModel}
- */
-class AvantGearData extends foundry.abstract.DataModel {
-    static defineSchema() {
-        const fields = foundry.data.fields;
-        return {
-            description: new fields.HTMLField({
-                required: true,
-                blank: true,
-                initial: ""
-            }),
-            weight: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 1,
-                min: 0
-            }),
-            cost: new fields.NumberField({
-                required: true,
-                nullable: false,
-                initial: 0,
-                min: 0
-            }),
-            quantity: new fields.NumberField({
-                required: true,
-                nullable: false,
-                integer: true,
-                initial: 1,
-                min: 0
-            }),
-            properties: new fields.StringField({
-                required: true,
-                blank: true,
-                initial: ""
-            })
-        };
-    }
-}
-
-/**
- * Actor Sheet - v12/v13 Compatible
- */
-class AvantActorSheet extends (foundry.appv1?.sheets?.ActorSheet || ActorSheet) {
-    static get defaultOptions() {
-        return foundry.utils.mergeObject(super.defaultOptions, {
-            classes: ["avant", "sheet", "actor"],
-            template: "systems/avant/templates/actor-sheet.html",
-            width: 900,
-            height: 630,
-            tabs: [{ 
-                navSelector: ".sheet-tabs", 
-                contentSelector: ".sheet-body", 
-                initial: "core"
-            }]
-        });
-    }
-
-    getData() {
-        const context = super.getData();
-        const actorData = this.actor.toObject(false);
+    prepareDerivedData() {
+        super.prepareDerivedData();
         
-        context.system = actorData.system;
-        context.flags = actorData.flags;
-        
-        // Calculate ability modifiers from ability scores (D&D style: (score-10)/2)
-        if (context.system.abilities) {
-            for (const [abilityName, abilityData] of Object.entries(context.system.abilities)) {
-                if (abilityData && typeof abilityData.value === 'number') {
-                    context.system.abilities[abilityName].mod = Math.floor((abilityData.value - 10) / 2);
-                }
-            }
+        // Call the data model's prepare derived data if it exists
+        if (this.system && typeof this.system.prepareDerivedData === 'function') {
+            this.system.prepareDerivedData();
         }
-        
-        // Prepare items by type for organized tabs
-        context.items = {};
-        for (let item of this.actor.items) {
-            const itemType = item.type;
-            if (!context.items[itemType]) context.items[itemType] = [];
-            context.items[itemType].push(item);
-        }
-        
-        // Ensure all item types exist even if empty
-        const itemTypes = ['action', 'feature', 'talent', 'augment', 'weapon', 'armor', 'gear'];
-        itemTypes.forEach(type => {
-            if (!context.items[type]) context.items[type] = [];
-        });
-        
-        // Dynamically organize skills by ability for template rendering
-        const skillAbilities = AvantActorData.getSkillAbilities();
-        context.skillsByAbility = {
-            might: [],
-            grace: [],
-            intellect: [],
-            focus: []
-        };
-        
-        // Group skills by their abilities
-        for (const [skillName, abilityName] of Object.entries(skillAbilities)) {
-            if (context.skillsByAbility[abilityName]) {
-                context.skillsByAbility[abilityName].push({
-                    name: skillName,
-                    label: skillName.charAt(0).toUpperCase() + skillName.slice(1),
-                    value: actorData.system.skills[skillName] || 0
-                });
-            }
-        }
-        
-        // Sort skills within each ability group alphabetically
-        Object.keys(context.skillsByAbility).forEach(ability => {
-            context.skillsByAbility[ability].sort((a, b) => a.label.localeCompare(b.label));
-        });
-        
-        return context;
-    }
-
-    _activateCoreListeners(html) {
-        // FoundryVTT v13 compatibility fix for core listeners
-        // Handle various types of HTML input that FoundryVTT might pass
-        let element = html;
-        let originalWasJQuery = html instanceof jQuery;
-        
-        // Handle jQuery objects by extracting the DOM element
-        if (html instanceof jQuery) {
-            if (html.length > 0) {
-                element = html[0];
-            } else {
-                console.error('AvantActorSheet._activateCoreListeners: Empty jQuery object received', html);
-                return;
-            }
-        }
-        
-        // Handle comment nodes or other non-element nodes
-        if (element && element.nodeType === Node.COMMENT_NODE) {
-            console.warn('AvantActorSheet._activateCoreListeners: Received comment node, looking for next element');
-            // Try to find the next element sibling
-            element = element.nextElementSibling;
-        }
-        
-        // Handle document fragments or other container types
-        if (element && element.nodeType === Node.DOCUMENT_FRAGMENT_NODE) {
-            // Look for the first element child
-            element = element.querySelector('form') || element.firstElementChild;
-        }
-        
-        // Final validation - ensure we have a valid DOM element
-        if (!element || !element.querySelectorAll || typeof element.querySelectorAll !== 'function') {
-            console.error('AvantActorSheet._activateCoreListeners: Could not find valid DOM element', html);
-            return;
-        }
-        
-        // FoundryVTT core expects jQuery objects, so wrap the DOM element back into jQuery
-        const jQueryElement = $(element);
-        
-        // Call parent with jQuery-wrapped element
-        super._activateCoreListeners(jQueryElement);
-    }
-
-    activateListeners(html) {
-        // FoundryVTT v13 ApplicationV1 compatibility
-        // Handle various types of HTML input that FoundryVTT might pass
-        let element = html;
-        
-        // Handle jQuery objects by extracting the DOM element
-        if (html instanceof jQuery) {
-            if (html.length > 0) {
-                element = html[0];
-            } else {
-                console.error('AvantActorSheet.activateListeners: Empty jQuery object received', html);
-                return;
-            }
-        }
-        
-        // Handle comment nodes or other non-element nodes
-        if (element && element.nodeType === Node.COMMENT_NODE) {
-            console.warn('AvantActorSheet.activateListeners: Received comment node, looking for next element');
-            // Try to find the next element sibling
-            element = element.nextElementSibling;
-        }
-        
-        // Handle document fragments or other container types
-        if (element && element.nodeType === Node.DOCUMENT_FRAGMENT_NODE) {
-            // Look for the first element child
-            element = element.querySelector('form') || element.firstElementChild;
-        }
-        
-        // Final validation - ensure we have a valid DOM element
-        if (!element || !element.querySelectorAll || typeof element.querySelectorAll !== 'function') {
-            console.error('AvantActorSheet.activateListeners: Could not find valid DOM element', html);
-            return;
-        }
-        
-        // FoundryVTT core expects jQuery objects, so wrap the DOM element back into jQuery
-        const jQueryElement = $(element);
-        
-        // Call parent with jQuery-wrapped element
-        super.activateListeners(jQueryElement);
-
-        // Everything below here is only needed if the sheet is editable
-        if (!this.isEditable) return;
-
-        // NOTE: Tab handling is done automatically by ApplicationV1 framework
-        // via the tabs configuration in defaultOptions. No manual initialization needed.
-
-        // Use pure DOM methods instead of jQuery for v13 compatibility
-        // Add Inventory Item
-        element.querySelectorAll('.item-create').forEach(el => {
-            el.addEventListener('click', this._onItemCreate.bind(this));
-        });
-
-        // Update Inventory Item
-        element.querySelectorAll('.item-edit').forEach(el => {
-            el.addEventListener('click', this._onItemEdit.bind(this));
-        });
-
-        // Delete Inventory Item
-        element.querySelectorAll('.item-delete').forEach(el => {
-            el.addEventListener('click', this._onItemDelete.bind(this));
-        });
-
-        // Rollable abilities
-        element.querySelectorAll('.rollable').forEach(el => {
-            el.addEventListener('click', this._onRoll.bind(this));
-        });
-
-        // Skill rolls
-        element.querySelectorAll('.skill-roll').forEach(el => {
-            el.addEventListener('click', this._onSkillRoll.bind(this));
-        });
-
-        // Power point usage
-        element.querySelectorAll('.power-point-use').forEach(el => {
-            el.addEventListener('click', this._onPowerPointsRoll.bind(this));
-        });
-
-        // Weapon attack rolls
-        element.querySelectorAll('.attack-roll').forEach(el => {
-            el.addEventListener('click', this._onAttackRoll.bind(this));
-        });
-
-        // Weapon damage rolls
-        element.querySelectorAll('.damage-roll').forEach(el => {
-            el.addEventListener('click', this._onDamageRoll.bind(this));
-        });
-
-        // Armor rolls
-        element.querySelectorAll('.armor-roll').forEach(el => {
-            el.addEventListener('click', this._onArmorRoll.bind(this));
-        });
-    }
-
-    async _onItemRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-
-        if (dataset.rollType) {
-            if (dataset.rollType == 'item') {
-                const itemElement = element.closest('.item');
-                if (!itemElement) {
-                    console.warn('Avant | No item element found for item roll');
-                    return;
-                }
-                
-                const itemId = itemElement.dataset.itemId;
-                const item = this.actor.items.get(itemId);
-                
-                if (!item) {
-                    console.warn(`Avant | Item with ID '${itemId}' not found`);
-                    ui.notifications.warn('Item not found');
-                    return;
-                }
-                
-                // Check if item has a roll method
-                if (typeof item.roll === 'function') {
-                    try {
-                        return await item.roll();
-                    } catch (error) {
-                        console.error('Avant | Error in item roll:', error);
-                        ui.notifications.error(`Failed to roll item: ${error.message}`);
-                    }
-                } else {
-                    // Create a basic roll for the item
-                    try {
-                        const roll = new Roll("1d20", {});
-                        await roll.evaluate();
-                        
-                        await roll.toMessage({
-                            speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                            flavor: `${item.name} Roll`,
-                            rollMode: game.settings.get('core', 'rollMode'),
-                        });
-                        return roll;
-                    } catch (error) {
-                        console.error('Avant | Error in basic item roll:', error);
-                        ui.notifications.error(`Failed to roll item: ${error.message}`);
-                    }
-                }
-            }
-        }
-
-        if (dataset.roll) {
-            try {
-                let label = dataset.label ? `[item] ${dataset.label}` : '';
-                let roll = new Roll(dataset.roll, this.actor.getRollData());
-                await roll.evaluate();
-                
-                await roll.toMessage({
-                    speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                    flavor: label,
-                    rollMode: game.settings.get('core', 'rollMode'),
-                });
-                return roll;
-            } catch (error) {
-                console.error('Avant | Error in dataset roll:', error);
-                ui.notifications.error(`Roll failed: ${error.message}`);
-            }
-        }
-    }
-
-    async _onAbilityRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const ability = dataset.ability;
-        
-        if (!ability) {
-            console.warn('Avant | No ability specified for roll');
-            return;
-        }
-        
-        const actor = this.actor;
-        const abilityData = actor.system.abilities[ability];
-        
-        if (!abilityData) {
-            console.warn(`Avant | Ability '${ability}' not found on actor`);
-            return;
-        }
-
-        // Check if this is a generation roll (Shift+Click) or a check roll (normal click)
-        const isGenerationRoll = event.shiftKey;
-        
-        try {
-            let roll, flavorText;
-            
-            if (isGenerationRoll) {
-                // Ability Score Generation: 4d6 drop lowest
-                roll = new Roll("4d6kh3");
-                await roll.evaluate();
-                
-                // Update the actor's ability score with the new value
-                const newValue = roll.total;
-                const updatePath = `system.abilities.${ability}.value`;
-                await actor.update({
-                    [updatePath]: newValue
-                });
-                
-                flavorText = `${ability.charAt(0).toUpperCase() + ability.slice(1)} Score Generation`;
-                ui.notifications.info(`Generated ${ability} score: ${newValue}`);
-            } else {
-                // Ability Check: 2d10 + Level + Ability Modifier (Avant system)
-                // Calculate ability modifier from ability score (standard D&D-style: (score-10)/2)
-                const abilityMod = Math.floor((abilityData.value - 10) / 2);
-                
-                roll = new Roll("2d10 + @level + @abilityMod", { 
-                    level: actor.system.level,
-                    abilityMod: abilityMod
-                });
-                await roll.evaluate();
-                
-                flavorText = `${ability.charAt(0).toUpperCase() + ability.slice(1)} Check`;
-            }
-            
-            await roll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: actor }),
-                flavor: flavorText,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            
-            return roll;
-        } catch (error) {
-            console.error('Avant | Error in ability roll:', error);
-            ui.notifications.error(`Failed to roll ${ability} check: ${error.message}`);
-        }
-    }
-
-    async _onItemCreate(event) {
-        event.preventDefault();
-        const header = event.currentTarget;
-        const type = header.dataset.type;
-        const data = foundry.utils.duplicate(header.dataset);
-        const name = `New ${type.capitalize()}`;
-        const itemData = {
-            name: name,
-            type: type,
-            system: data
-        };
-        delete itemData.system["type"];
-        
-        // Handle specific item type defaults
-        if (type === "feature" && data.category) {
-            itemData.system.category = data.category;
-        }
-        if (type === "action" && !itemData.system.ability) {
-            itemData.system.ability = "might";
-        }
-        if (type === "augment" && !itemData.system.augmentType) {
-            itemData.system.augmentType = "enhancement";
-        }
-        
-        try {
-            return await Item.create(itemData, {parent: this.actor});
-        } catch (error) {
-            console.error('Avant | Error creating item:', error);
-            ui.notifications.error(`Failed to create ${type}: ${error.message}`);
-        }
-    }
-
-    _onItemEdit(event) {
-        event.preventDefault();
-        const li = event.currentTarget.closest(".item");
-        if (!li) {
-            console.warn('Avant | No item element found for edit');
-            return;
-        }
-        
-        const itemId = li.dataset.itemId;
-        const item = this.actor.items.get(itemId);
-        
-        if (!item) {
-            console.warn(`Avant | Item with ID '${itemId}' not found`);
-            ui.notifications.warn('Item not found');
-            return;
-        }
-        
-        item.sheet.render(true);
-    }
-
-    async _onItemDelete(event) {
-        event.preventDefault();
-        const li = event.currentTarget.closest(".item");
-        if (!li) {
-            console.warn('Avant | No item element found for delete');
-            return;
-        }
-        
-        const itemId = li.dataset.itemId;
-        const item = this.actor.items.get(itemId);
-        
-        if (!item) {
-            console.warn(`Avant | Item with ID '${itemId}' not found`);
-            ui.notifications.warn('Item not found');
-            return;
-        }
-        
-        try {
-            await item.delete();
-            // FoundryVTT v13 compatibility: Convert DOM element to jQuery object for slideUp
-            const $li = $(li);
-            if (typeof $li.slideUp === 'function') {
-                $li.slideUp(200, () => this.render(false));
-            } else {
-                // Fallback: Remove element immediately and re-render
-                li.remove();
-                this.render(false);
-            }
-        } catch (error) {
-            console.error('Avant | Error deleting item:', error);
-            ui.notifications.error(`Failed to delete item: ${error.message}`);
-        }
-    }
-
-    async _onRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-
-        // Handle different types of rolls
-        if (dataset.rollType) {
-            if (dataset.rollType === 'ability') {
-                return await this._onAbilityRoll(event);
-            }
-            if (dataset.rollType === 'item') {
-                return await this._onItemRoll(event);
-            }
-        }
-
-        // Handle generic rolls
-        if (dataset.roll) {
-            try {
-                const roll = new Roll(dataset.roll, this.actor.getRollData());
-                await roll.evaluate();
-                
-                const label = dataset.label ? `${dataset.label}` : '';
-                await roll.toMessage({
-                    speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                    flavor: label,
-                    rollMode: game.settings.get('core', 'rollMode'),
-                });
-                return roll;
-            } catch (error) {
-                console.error('Avant | Error in generic roll:', error);
-                ui.notifications.error(`Roll failed: ${error.message}`);
-            }
-        }
-    }
-
-    async _onSkillRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const skill = dataset.skill;
-        
-        if (!skill) {
-            console.warn('Avant | No skill specified for roll');
-            return;
-        }
-        
-        const actor = this.actor;
-        const skillValue = actor.system.skills[skill];
-        const skillAbilities = AvantActorData.getSkillAbilities();
-        const abilityKey = skillAbilities[skill];
-        const abilityData = actor.system.abilities[abilityKey];
-        
-        // Calculate ability modifier dynamically from ability score (standard D&D-style: (score-10)/2)
-        const abilityMod = abilityData ? Math.floor((abilityData.value - 10) / 2) : 0;
-        
-        if (skillValue === undefined) {
-            console.warn(`Avant | Skill '${skill}' not found on actor`);
-            return;
-        }
-        
-        try {
-            // Avant uses 2d10 + Level + Ability Modifier + Skill Modifier
-            const roll = new Roll("2d10 + @level + @abilityMod + @skillMod", {
-                level: actor.system.level,
-                abilityMod: abilityMod,
-                skillMod: skillValue
-            });
-            await roll.evaluate();
-            
-            const skillName = skill.charAt(0).toUpperCase() + skill.slice(1);
-            const abilityName = abilityKey.charAt(0).toUpperCase() + abilityKey.slice(1);
-            
-            await roll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: actor }),
-                flavor: `${skillName} Check (${abilityName})`,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            
-            return roll;
-        } catch (error) {
-            console.error('Avant | Error in skill roll:', error);
-            ui.notifications.error(`Failed to roll ${skill} check: ${error.message}`);
-        }
-    }
-
-    async _onPowerPointsRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const cost = parseInt(dataset.cost) || 1;
-        
-        const actor = this.actor;
-        const currentPP = actor.system.powerPoints.value;
-        
-        if (currentPP < cost) {
-            ui.notifications.warn(`Not enough Power Points! Need ${cost}, have ${currentPP}`);
-            return;
-        }
-        
-        // Deduct power points
-        await actor.update({
-            "system.powerPoints.value": Math.max(0, currentPP - cost)
-        });
-        
-        ui.notifications.info(`Spent ${cost} Power Point${cost > 1 ? 's' : ''}. Remaining: ${currentPP - cost}`);
-    }
-
-    async _onAttackRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const itemId = dataset.itemId;
-        
-        if (!itemId) {
-            console.warn('Avant | No item ID found for attack roll');
-            return;
-        }
-        
-        const weapon = this.actor.items.get(itemId);
-        
-        if (!weapon) {
-            console.warn(`Avant | Weapon with ID '${itemId}' not found`);
-            ui.notifications.warn('Weapon not found');
-            return;
-        }
-        
-        try {
-            // Use weapon's ability and modifier for attack rolls
-            const weaponAbility = weapon.system.ability || 'might';
-            const abilityMod = this.actor.system.abilities[weaponAbility]?.mod || 0;
-            const weaponModifier = weapon.system.modifier || 0;
-            
-            const roll = new Roll("2d10 + @level + @abilityMod + @weaponMod", {
-                level: this.actor.system.level,
-                abilityMod: abilityMod,
-                weaponMod: weaponModifier
-            });
-            await roll.evaluate();
-            
-            await roll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                flavor: `${weapon.name} Attack`,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            return roll;
-        } catch (error) {
-            console.error('Avant | Error in weapon attack roll:', error);
-            ui.notifications.error(`Failed to roll weapon attack: ${error.message}`);
-        }
-    }
-
-    async _onDamageRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const itemId = dataset.itemId;
-        
-        if (!itemId) {
-            console.warn('Avant | No item ID found for damage roll');
-            return;
-        }
-        
-        const weapon = this.actor.items.get(itemId);
-        
-        if (!weapon) {
-            console.warn(`Avant | Weapon with ID '${itemId}' not found`);
-            ui.notifications.warn('Weapon not found');
-            return;
-        }
-        
-        try {
-            // Use the weapon's damage dice and ability modifier
-            const damageRoll = weapon.system.damageDie || "1d6";
-            const weaponAbility = weapon.system.ability || 'might';
-            const abilityMod = this.actor.system.abilities[weaponAbility]?.mod || 0;
-            const damageType = weapon.system.damageType || "";
-            
-            const roll = new Roll(`${damageRoll} + @abilityMod`, {
-                abilityMod: abilityMod
-            });
-            await roll.evaluate();
-            
-            const flavorText = damageType ? 
-                `${weapon.name} Damage (${damageType})` : 
-                `${weapon.name} Damage`;
-            
-            await roll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                flavor: flavorText,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            return roll;
-        } catch (error) {
-            console.error('Avant | Error in weapon damage roll:', error);
-            ui.notifications.error(`Failed to roll weapon damage: ${error.message}`);
-        }
-    }
-
-    async _onArmorRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-        const itemId = dataset.itemId;
-        
-        if (!itemId) {
-            console.warn('Avant | No item ID found for armor roll');
-            return;
-        }
-        
-        const armor = this.actor.items.get(itemId);
-        
-        if (!armor) {
-            console.warn(`Avant | Armor with ID '${itemId}' not found`);
-            ui.notifications.warn('Armor not found');
-            return;
-        }
-        
-        try {
-            // Use armor's ability and modifier for armor rolls
-            const armorAbility = armor.system.ability || 'grace';
-            const abilityMod = this.actor.system.abilities[armorAbility]?.mod || 0;
-            const armorModifier = armor.system.modifier || 0;
-            
-            const roll = new Roll("2d10 + @level + @abilityMod + @armorMod", {
-                level: this.actor.system.level,
-                abilityMod: abilityMod,
-                armorMod: armorModifier
-            });
-            await roll.evaluate();
-            
-            await roll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                flavor: `${armor.name} Armor Check`,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            return roll;
-        } catch (error) {
-            console.error('Avant | Error in armor roll:', error);
-            ui.notifications.error(`Failed to roll armor check: ${error.message}`);
-        }
-    }
-}
-
-/**
- * Item Sheet - v12/v13 Compatible
- */
-class AvantItemSheet extends (foundry.appv1?.sheets?.ItemSheet || ItemSheet) {
-    static get defaultOptions() {
-        return foundry.utils.mergeObject(super.defaultOptions, {
-            classes: ["avant", "sheet", "item"],
-            width: 520,
-            height: 480,
-            tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }]
-        });
-    }
-
-    get template() {
-        const path = "systems/avant/templates/item";
-        return `${path}/item-${this.item.type}-sheet.html`;
-    }
-
-    getData() {
-        const context = super.getData();
-        const itemData = this.item.toObject(false);
-        context.system = itemData.system;
-        context.flags = itemData.flags;
-        return context;
-    }
-
-    _activateCoreListeners(html) {
-        // FoundryVTT v13 compatibility fix for core listeners
-        // Handle various types of HTML input that FoundryVTT might pass
-        let element = html;
-        
-        // Handle jQuery objects by extracting the DOM element
-        if (html instanceof jQuery) {
-            if (html.length > 0) {
-                element = html[0];
-            } else {
-                console.error('AvantItemSheet._activateCoreListeners: Empty jQuery object received', html);
-                return;
-            }
-        }
-        
-        // Handle comment nodes or other non-element nodes
-        if (element && element.nodeType === Node.COMMENT_NODE) {
-            console.warn('AvantItemSheet._activateCoreListeners: Received comment node, looking for next element');
-            // Try to find the next element sibling
-            element = element.nextElementSibling;
-        }
-        
-        // Handle document fragments or other container types
-        if (element && element.nodeType === Node.DOCUMENT_FRAGMENT_NODE) {
-            // Look for the first element child
-            element = element.querySelector('form') || element.firstElementChild;
-        }
-        
-        // Final validation - ensure we have a valid DOM element
-        if (!element || !element.querySelectorAll || typeof element.querySelectorAll !== 'function') {
-            console.error('AvantItemSheet._activateCoreListeners: Could not find valid DOM element', html);
-            return;
-        }
-        
-        // FoundryVTT core expects jQuery objects, so wrap the DOM element back into jQuery
-        const jQueryElement = $(element);
-        
-        // Call parent with jQuery-wrapped element
-        super._activateCoreListeners(jQueryElement);
-    }
-
-    activateListeners(html) {
-        // FoundryVTT v13 ApplicationV1 compatibility
-        // Handle various types of HTML input that FoundryVTT might pass
-        let element = html;
-        
-        // Handle jQuery objects by extracting the DOM element
-        if (html instanceof jQuery) {
-            if (html.length > 0) {
-                element = html[0];
-            } else {
-                console.error('AvantItemSheet.activateListeners: Empty jQuery object received', html);
-                return;
-            }
-        }
-        
-        // Handle comment nodes or other non-element nodes
-        if (element && element.nodeType === Node.COMMENT_NODE) {
-            console.warn('AvantItemSheet.activateListeners: Received comment node, looking for next element');
-            // Try to find the next element sibling
-            element = element.nextElementSibling;
-        }
-        
-        // Handle document fragments or other container types
-        if (element && element.nodeType === Node.DOCUMENT_FRAGMENT_NODE) {
-            // Look for the first element child
-            element = element.querySelector('form') || element.firstElementChild;
-        }
-        
-        // Final validation - ensure we have a valid DOM element
-        if (!element || !element.querySelectorAll || typeof element.querySelectorAll !== 'function') {
-            console.error('AvantItemSheet.activateListeners: Could not find valid DOM element', html);
-            return;
-        }
-        
-        // FoundryVTT core expects jQuery objects, so wrap the DOM element back into jQuery
-        const jQueryElement = $(element);
-        
-        // Call parent with jQuery-wrapped element
-        super.activateListeners(jQueryElement);
-
-        if (!this.isEditable) return;
-
-        // Use pure DOM methods instead of jQuery for v13 compatibility
-        // Rollable abilities
-        element.querySelectorAll('.rollable').forEach(el => {
-            el.addEventListener('click', this._onRoll.bind(this));
-        });
-    }
-
-    async _onRoll(event) {
-        event.preventDefault();
-        const element = event.currentTarget;
-        const dataset = element.dataset;
-
-        if (dataset.roll) {
-            try {
-                const roll = new Roll(dataset.roll, this.item.getRollData());
-                await roll.evaluate();
-                
-                const label = dataset.label ? `${dataset.label}` : this.item.name;
-                await roll.toMessage({
-                    speaker: ChatMessage.getSpeaker({ actor: this.item.actor }),
-                    flavor: label,
-                    rollMode: game.settings.get('core', 'rollMode'),
-                });
-                return roll;
-            } catch (error) {
-                console.error('Avant | Error in item sheet roll:', error);
-                ui.notifications.error(`Roll failed: ${error.message}`);
-            }
-        }
-    }
-
-    /**
-     * Override the default update object method to ensure proper data type conversion
-     * @param {Event} event - The form submission event
-     * @param {Object} formData - The form data to process
-     * @returns {Promise<Object>} The processed update data
-     */
-    async _updateObject(event, formData) {
-        // Convert string values to numbers for specific fields that should be integers
-        const processedData = foundry.utils.expandObject(formData);
-        
-        // Handle talent and augment specific fields
-        if (processedData.system) {
-            // Convert powerPointCost to integer if present
-            if (processedData.system.powerPointCost !== undefined) {
-                processedData.system.powerPointCost = parseInt(processedData.system.powerPointCost) || 0;
-            }
-            
-            // Convert uses fields to integers if present
-            if (processedData.system.uses) {
-                if (processedData.system.uses.value !== undefined) {
-                    processedData.system.uses.value = parseInt(processedData.system.uses.value) || 0;
-                }
-                if (processedData.system.uses.max !== undefined) {
-                    processedData.system.uses.max = parseInt(processedData.system.uses.max) || 0;
-                }
-            }
-            
-            // Handle weapon/armor specific integer fields
-            if (processedData.system.modifier !== undefined) {
-                processedData.system.modifier = parseInt(processedData.system.modifier) || 0;
-            }
-            if (processedData.system.threshold !== undefined) {
-                processedData.system.threshold = parseInt(processedData.system.threshold) || 11;
-            }
-            if (processedData.system.armorClass !== undefined) {
-                processedData.system.armorClass = parseInt(processedData.system.armorClass) || 10;
-            }
-            
-            // Handle numeric fields that should be numbers (not necessarily integers)
-            if (processedData.system.weight !== undefined) {
-                processedData.system.weight = parseFloat(processedData.system.weight) || 0;
-            }
-            if (processedData.system.cost !== undefined) {
-                processedData.system.cost = parseFloat(processedData.system.cost) || 0;
-            }
-        }
-        
-        // Convert back to flat object for FoundryVTT
-        const flatData = foundry.utils.flattenObject(processedData);
-        
-        // Call parent method with processed data
-        return super._updateObject(event, flatData);
-    }
-}
-
-/**
- * Fortune Point Reroll Dialog
- * Allows players to reroll individual d10s from their previous roll
- * @extends {Application}
- */
-class AvantRerollDialog extends Application {
-    constructor(originalRoll, actor, originalFlavor, options = {}) {
-        super(options);
-        this.originalRoll = originalRoll;
-        this.actor = actor;
-        this.originalFlavor = originalFlavor;
-        this.selectedDice = new Set();
-        
-        // Extract d10 dice results from the original roll
-        this.d10Results = this._extractD10Results(originalRoll);
-        this.staticModifiers = this._extractStaticModifiers(originalRoll);
-    }
-    
-    static get defaultOptions() {
-        return foundry.utils.mergeObject(super.defaultOptions, {
-            id: "avant-reroll-dialog",
-            classes: ["avant", "dialog", "reroll-dialog"],
-            title: "Fortune Point Reroll",
-            template: "systems/avant/templates/reroll-dialog.html",
-            width: 400,
-            height: 300,
-            resizable: false
-        });
-    }
-    
-    getData() {
-        const fortunePoints = this.actor.system.fortunePoints || 0;
-        const maxRerolls = Math.min(this.d10Results.length, fortunePoints);
-        
-        return {
-            d10Results: this.d10Results.map((result, index) => ({
-                index: index,
-                value: result.value,
-                selected: this.selectedDice.has(index),
-                wasRerolled: result.wasRerolled,
-                canReroll: !result.wasRerolled // Cannot reroll if already rerolled
-            })),
-            selectedCount: this.selectedDice.size,
-            fortunePoints: fortunePoints,
-            maxRerolls: maxRerolls,
-            canReroll: fortunePoints > 0 && this.selectedDice.size > 0,
-            costMessage: this._getCostMessage(fortunePoints),
-            originalTotal: this.originalRoll.total,
-            originalFlavor: this.originalFlavor
-        };
-    }
-    
-    activateListeners(html) {
-        super.activateListeners(html);
-        
-        // Dice selection
-        html.find('.reroll-die').click(this._onDieClick.bind(this));
-        
-        // Reroll button
-        html.find('.reroll-confirm').click(this._onRerollConfirm.bind(this));
-        
-        // Cancel button
-        html.find('.reroll-cancel').click(this._onCancel.bind(this));
-    }
-    
-    async _onDieClick(event) {
-        const dieIndex = parseInt(event.currentTarget.dataset.index);
-        const fortunePoints = this.actor.system.fortunePoints || 0;
-        const dieData = this.d10Results[dieIndex];
-        
-        // Prevent selection of already rerolled dice
-        if (dieData.wasRerolled) {
-            ui.notifications.warn("This die has already been rerolled and cannot be rerolled again!");
-            return;
-        }
-        
-        if (this.selectedDice.has(dieIndex)) {
-            // Deselect die
-            this.selectedDice.delete(dieIndex);
-        } else {
-            // Select die if we have enough Fortune Points
-            if (this.selectedDice.size < fortunePoints) {
-                this.selectedDice.add(dieIndex);
-            } else {
-                ui.notifications.warn("Not enough Fortune Points to select more dice!");
-                return;
-            }
-        }
-        
-        this.render(false);
-    }
-    
-    async _onRerollConfirm(event) {
-        const fortunePoints = this.actor.system.fortunePoints || 0;
-        const selectedCount = this.selectedDice.size;
-        
-        if (selectedCount === 0) {
-            ui.notifications.warn("Please select at least one die to reroll.");
-            return;
-        }
-        
-        if (selectedCount > fortunePoints) {
-            ui.notifications.warn("Not enough Fortune Points!");
-            return;
-        }
-        
-        try {
-            // Create new roll with rerolled dice
-            const newRoll = await this._createReroll();
-            
-            // Deduct Fortune Points
-            const newFortunePoints = Math.max(0, fortunePoints - selectedCount);
-            await this.actor.update({
-                "system.fortunePoints": newFortunePoints
-            });
-            
-            // Post new roll to chat
-            await newRoll.toMessage({
-                speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-                flavor: `${this.originalFlavor} (Rerolled with ${selectedCount} Fortune Point${selectedCount > 1 ? 's' : ''})`,
-                rollMode: game.settings.get('core', 'rollMode'),
-            });
-            
-            ui.notifications.info(`Rerolled ${selectedCount} die with Fortune Points. Remaining: ${newFortunePoints}`);
-            this.close();
-            
-        } catch (error) {
-            console.error('Avant | Error in reroll:', error);
-            ui.notifications.error(`Reroll failed: ${error.message}`);
-        }
-    }
-    
-    _onCancel(event) {
-        this.close();
     }
     
     /**
-     * Extract d10 results from the original roll
-     * @param {Roll} roll - The original roll
-     * @returns {Array} Array of d10 die data with values and reroll status
+     * Get roll data for this actor
+     * @returns {Object} Roll data for use in roll formulas
+     * @override
      */
-    _extractD10Results(roll) {
-        const d10Results = [];
+    getRollData() {
+        const data = super.getRollData();
         
-        // Navigate through the roll terms to find d10 dice
-        for (const term of roll.terms) {
-            if (term instanceof foundry.dice.terms.Die && term.faces === 10) {
-                // Extract individual die results with reroll status
-                for (const result of term.results) {
-                    if (result.active) {
-                        d10Results.push({
-                            value: result.result,
-                            wasRerolled: result.rerolled || false
-                        });
-                    }
-                }
+        // Add commonly used roll data
+        if (this.system.abilities) {
+            for (const [abilityName, abilityData] of Object.entries(this.system.abilities)) {
+                data[abilityName] = abilityData.value;
+                data[`${abilityName}Mod`] = abilityData.mod;
             }
         }
         
-        return d10Results;
-    }
-    
-    /**
-     * Extract static modifiers from the original roll (level, ability mod, etc.)
-     * @param {Roll} roll - The original roll
-     * @returns {number} Total static modifiers
-     */
-    _extractStaticModifiers(roll) {
-        let staticModifiers = 0;
-        
-        for (const term of roll.terms) {
-            if (term instanceof foundry.dice.terms.NumericTerm) {
-                staticModifiers += term.number;
+        if (this.system.skills) {
+            for (const [skillName, skillValue] of Object.entries(this.system.skills)) {
+                data[skillName] = skillValue;
             }
         }
         
-        return staticModifiers;
+        data.level = this.system.level || 1;
+        data.tier = this.system.tier || 1;
+        data.effort = this.system.effort || 1;
+        
+        return data;
+    }
+}
+
+/**
+ * Extended Item class for Avant Native
+ * @extends {Item}
+ */
+class AvantItem extends Item {
+    /**
+     * Prepare item data
+     * Validate and normalize item data
+     * @override
+     */
+    prepareData() {
+        super.prepareData();
+        
+        // Validate item data
+        this.system = ValidationUtils.validateItemData(this.system);
     }
     
     /**
-     * Create a new roll with rerolled dice
-     * @returns {Roll} New roll with rerolled dice
+     * Get roll data for this item
+     * @returns {Object} Roll data for use in roll formulas
+     * @override
      */
-    async _createReroll() {
-        const newD10Results = [...this.d10Results];
+    getRollData() {
+        const data = super.getRollData();
         
-        // Reroll selected dice
-        for (const dieIndex of this.selectedDice) {
-            const newRoll = await new Roll("1d10").evaluate();
-            newD10Results[dieIndex] = {
-                value: newRoll.total,
-                wasRerolled: true // Mark this die as rerolled
-            };
+        // Add actor data if this item is owned
+        if (this.actor) {
+            const actorData = this.actor.getRollData();
+            Object.assign(data, actorData);
         }
         
-        // Calculate new total
-        const diceTotal = newD10Results.reduce((sum, die) => sum + die.value, 0);
-        const newTotal = diceTotal + this.staticModifiers;
+        return data;
+    }
+    
+    /**
+     * Handle item rolls
+     * @returns {Promise<Roll|void>} The executed roll or void
+     */
+    async roll() {
+        // Default item roll behavior
+        const roll = new Roll("1d20", this.getRollData());
+        await roll.evaluate();
         
-        // Create a new Roll object with the new results
-        // We'll create a simple roll that represents the final result
-        const rollFormula = `${newD10Results.map(die => die.value).join(' + ')} + ${this.staticModifiers}`;
-        const roll = new Roll(rollFormula);
-        
-        // Manually set the evaluated state and total
-        roll._evaluated = true;
-        roll._total = newTotal;
-        
-        // Create the roll terms manually to show individual dice
-        const diceTerms = [];
-        for (let i = 0; i < newD10Results.length; i++) {
-            if (i > 0) diceTerms.push(new foundry.dice.terms.OperatorTerm({ operator: '+' }));
-            
-            const dieTerm = new foundry.dice.terms.Die({ 
-                number: 1, 
-                faces: 10 
-            });
-            dieTerm.results = [{
-                result: newD10Results[i].value,
-                active: true,
-                rerolled: newD10Results[i].wasRerolled
-            }];
-            dieTerm._evaluated = true;
-            diceTerms.push(dieTerm);
-        }
-        
-        if (this.staticModifiers !== 0) {
-            diceTerms.push(new foundry.dice.terms.OperatorTerm({ operator: '+' }));
-            diceTerms.push(new foundry.dice.terms.NumericTerm({ number: this.staticModifiers }));
-        }
-        
-        roll.terms = diceTerms;
+        await roll.toMessage({
+            speaker: ChatMessage.getSpeaker({ actor: this.actor }),
+            flavor: `${this.name} Roll`,
+            rollMode: game.settings.get('core', 'rollMode'),
+        });
         
         return roll;
     }
-    
-    /**
-     * Generate cost message based on current Fortune Points
-     * @param {number} fortunePoints - Current Fortune Points
-     * @returns {string} Cost message
-     */
-    _getCostMessage(fortunePoints) {
-        if (fortunePoints === 0) {
-            return "No Fortune Points available!";
-        }
-        
-        const selectedCount = this.selectedDice.size;
-        if (selectedCount === 0) {
-            return `Select dice to reroll (${fortunePoints} Fortune Point${fortunePoints > 1 ? 's' : ''} available)`;
-        }
-        
-        return `Reroll for ${selectedCount} Fortune Point${selectedCount > 1 ? 's' : ''}`;
-    }
 }
 
-/**
- * Chat Message Context Menu Handler - Version-Aware Implementation
- * Supports both v12 and v13 FoundryVTT versions
- */
-class AvantChatContextMenu {
-    static addContextMenuListeners() {
-        console.log('Avant | Initializing version-aware context menu system...');
-        
-        // Detect Foundry version
-        const foundryVersion = game.version || game.data.version;
-        const isMajorV13 = foundryVersion && foundryVersion.startsWith('13');
-        const isMajorV12 = foundryVersion && foundryVersion.startsWith('12');
-        
-        console.log(`Avant | Detected Foundry version: ${foundryVersion}, Major v13: ${isMajorV13}, Major v12: ${isMajorV12}`);
-        
-        if (isMajorV13) {
-            console.log('Avant | Using v13 approach: Direct ChatLog._getEntryContextOptions extension');
-            AvantChatContextMenu._initializeV13Approach();
-        } else if (isMajorV12) {
-            console.log('Avant | Using v12 approach: Traditional hook-based method');
-            AvantChatContextMenu._initializeV12Approach();
-        } else {
-            console.log('Avant | Unknown version, attempting v12 approach as fallback');
-            AvantChatContextMenu._initializeV12Approach();
-        }
-        
-        console.log('Avant | Context menu listeners registered successfully');
-    }
-    
-    /**
-     * Initialize context menu for FoundryVTT v13
-     * Uses direct ChatLog method extension (PROVEN WORKING)
-     */
-    static _initializeV13Approach() {
-        // Wait for ChatLog to be available
-        if (ui.chat) {
-            AvantChatContextMenu._extendChatLogContextMenuV13();
-        } else {
-            // Wait for UI to be ready
-            Hooks.once('ready', () => {
-                AvantChatContextMenu._extendChatLogContextMenuV13();
-            });
-        }
-    }
-    
-    /**
-     * Initialize context menu for FoundryVTT v12
-     * Uses traditional hook-based approach
-     */
-    static _initializeV12Approach() {
-        console.log('Avant | Setting up v12 hook-based context menu...');
-        
-        // Try getChatLogEntryContext hook first (most common in v12)
-        Hooks.on('getChatLogEntryContext', (html, options) => {
-            console.log('Avant | 🎯 getChatLogEntryContext hook fired (v12)');
-            AvantChatContextMenu._addRerollOptionV12(html, options);
-        });
-        
-        // Also try getDocumentContextOptions as backup
-        Hooks.on('getDocumentContextOptions', (document, options) => {
-            console.log('Avant | 🎯 getDocumentContextOptions hook fired (v12)');
-            if (document instanceof ChatMessage) {
-                AvantChatContextMenu._addRerollOptionForDocument(document, options);
-            }
-        });
-        
-        console.log('Avant | v12 hooks registered');
-    }
-    
-    /**
-     * Extend ChatLog's context menu for v13 (WORKING SOLUTION)
-     */
-    static _extendChatLogContextMenuV13() {
-        console.log('Avant | 🎯 Extending ChatLog context menu for v13...');
-        
-        if (!ui.chat) {
-            console.log('Avant | ERROR: ui.chat not available');
-            return;
-        }
-        
-        // Store the original method
-        const originalGetEntryContextOptions = ui.chat._getEntryContextOptions;
-        
-        if (!originalGetEntryContextOptions) {
-            console.log('Avant | ERROR: _getEntryContextOptions method not found on ChatLog');
-            return;
-        }
-        
-        console.log('Avant | Found _getEntryContextOptions method, extending...');
-        
-        // Override the method
-        ui.chat._getEntryContextOptions = function() {
-            console.log('Avant | 🎯 EXTENDED _getEntryContextOptions called (v13)!');
-            
-            // Get the original options
-            const options = originalGetEntryContextOptions.call(this);
-            console.log('Avant | Original options:', options.map(opt => opt.name));
-            
-            // Add our reroll option
-            options.push({
-                name: "Reroll with Fortune Points",
-                icon: '<i class="fas fa-dice"></i>',
-                condition: (li) => {
-                    console.log('Avant | Condition check - li element (v13):', li);
-                    
-                    // v13 DOM compatibility: use vanilla DOM methods
-                    const messageId = li.dataset?.messageId || li.getAttribute('data-message-id');
-                    console.log('Avant | Message ID from li (v13):', messageId);
-                    
-                    if (!messageId) return false;
-                    
-                    const message = game.messages.get(messageId);
-                    console.log('Avant | Message object (v13):', message);
-                    
-                    if (!message || !message.rolls || message.rolls.length === 0) {
-                        console.log('Avant | No rolls in message (v13)');
-                        return false;
-                    }
-                    
-                    const roll = message.rolls[0];
-                    const actor = AvantChatContextMenu._getActorFromMessage(message);
-                    const isEligible = AvantChatContextMenu._isEligibleRoll(roll);
-                    
-                    console.log('Avant | Condition check result (v13) - eligible:', isEligible, 'actor:', !!actor);
-                    return isEligible && actor;
-                },
-                callback: (li) => {
-                    console.log('Avant | === CONTEXT MENU CALLBACK TRIGGERED (v13) ===');
-                    console.log('Avant | Callback li element (v13):', li);
-                    
-                    const messageId = li.dataset?.messageId || li.getAttribute('data-message-id');
-                    const message = game.messages.get(messageId);
-                    const roll = message.rolls[0];
-                    const actor = AvantChatContextMenu._getActorFromMessage(message);
-                    
-                    const dialog = new AvantRerollDialog(roll, actor, message.flavor);
-                    dialog.render(true);
-                    console.log('Avant | Dialog rendered (v13)');
-                }
-            });
-            
-            console.log('Avant | Extended options (v13):', options.map(opt => opt.name));
-            return options;
-        };
-        
-        console.log('Avant | ✅ ChatLog context menu extended successfully for v13!');
-    }
-    
-    /**
-     * Add reroll option for v12 hook-based approach
-     */
-    static _addRerollOptionV12(html, options) {
-        console.log('Avant | Adding reroll option via v12 hook approach');
-        
-        options.push({
-            name: "Reroll with Fortune Points",
-            icon: '<i class="fas fa-dice"></i>',
-            condition: (li) => {
-                console.log('Avant | Condition check - li element (v12):', li);
-                
-                // v12 jQuery compatibility: use jQuery methods if available
-                let messageId;
-                if (li.data && typeof li.data === 'function') {
-                    messageId = li.data('messageId') || li.attr('data-message-id');
-                } else {
-                    // Fallback to vanilla DOM
-                    messageId = li.dataset?.messageId || li.getAttribute('data-message-id');
-                }
-                
-                console.log('Avant | Message ID from li (v12):', messageId);
-                
-                if (!messageId) return false;
-                
-                const message = game.messages.get(messageId);
-                console.log('Avant | Message object (v12):', message);
-                
-                if (!message || !message.rolls || message.rolls.length === 0) {
-                    console.log('Avant | No rolls in message (v12)');
-                    return false;
-                }
-                
-                const roll = message.rolls[0];
-                const actor = AvantChatContextMenu._getActorFromMessage(message);
-                const isEligible = AvantChatContextMenu._isEligibleRoll(roll);
-                
-                console.log('Avant | Condition check result (v12) - eligible:', isEligible, 'actor:', !!actor);
-                return isEligible && actor;
-            },
-            callback: (li) => {
-                console.log('Avant | === CONTEXT MENU CALLBACK TRIGGERED (v12) ===');
-                console.log('Avant | Callback li element (v12):', li);
-                
-                let messageId;
-                if (li.data && typeof li.data === 'function') {
-                    messageId = li.data('messageId') || li.attr('data-message-id');
-                } else {
-                    messageId = li.dataset?.messageId || li.getAttribute('data-message-id');
-                }
-                
-                const message = game.messages.get(messageId);
-                const roll = message.rolls[0];
-                const actor = AvantChatContextMenu._getActorFromMessage(message);
-                
-                const dialog = new AvantRerollDialog(roll, actor, message.flavor);
-                dialog.render(true);
-                console.log('Avant | Dialog rendered (v12)');
-            }
-        });
-        
-        console.log('Avant | Reroll option added via v12 hook');
-    }
+// Expose classes globally for console access and module compatibility
+globalThis.AvantActorData = AvantActorData;
+globalThis.AvantActionData = AvantActionData;
+globalThis.AvantFeatureData = AvantFeatureData;
+globalThis.AvantTalentData = AvantTalentData;
+globalThis.AvantAugmentData = AvantAugmentData;
+globalThis.AvantWeaponData = AvantWeaponData;
+globalThis.AvantArmorData = AvantArmorData;
+globalThis.AvantGearData = AvantGearData;
+globalThis.AvantActorSheet = AvantActorSheet;
+globalThis.AvantItemSheet = AvantItemSheet;
+globalThis.AvantRerollDialog = AvantRerollDialog;
+globalThis.AvantChatContextMenu = AvantChatContextMenu;
+globalThis.AvantThemeManager = AvantThemeManager;
+globalThis.CompatibilityUtils = CompatibilityUtils;
+globalThis.ValidationUtils = ValidationUtils;
 
-    /**
-     * Add reroll option for document-based context menu (v12 fallback)
-     */
-    static _addRerollOptionForDocument(document, options) {
-        console.log('Avant | Adding reroll option for document approach (v12 fallback)');
-        
-        if (!(document instanceof ChatMessage)) {
-            console.log('Avant | Document is not a ChatMessage');
-            return;
-        }
-        
-        const message = document;
-        
-        if (!message.rolls || message.rolls.length === 0) {
-            console.log('Avant | No rolls in message (document approach)');
-            return;
-        }
-        
-        const roll = message.rolls[0];
-        const actor = AvantChatContextMenu._getActorFromMessage(message);
-        const isEligible = AvantChatContextMenu._isEligibleRoll(roll);
-        
-        if (!isEligible || !actor) {
-            console.log('Avant | Message not eligible for reroll (document approach)');
-            return;
-        }
-        
-        options.push({
-            name: "Reroll with Fortune Points",
-            icon: '<i class="fas fa-dice"></i>',
-            condition: () => true, // Already checked above
-            callback: () => {
-                console.log('Avant | === DOCUMENT CONTEXT MENU CALLBACK TRIGGERED ===');
-                const dialog = new AvantRerollDialog(roll, actor, message.flavor);
-                dialog.render(true);
-                console.log('Avant | Dialog rendered (document approach)');
-            }
-        });
-        
-        console.log('Avant | Reroll option added via document approach');
-    }
-    
-    /**
-     * Get actor from chat message
-     * @param {ChatMessage} message - The chat message
-     * @returns {Actor|null} The actor or null
-     */
-    static _getActorFromMessage(message) {
-        console.log('Avant | === _getActorFromMessage START ===');
-        console.log('Avant | Message ID:', message?.id);
-        console.log('Avant | Message speaker:', message?.speaker);
-        console.log('Avant | Speaker actor ID:', message?.speaker?.actor);
-        
-        if (message?.speaker?.actor) {
-            const actorId = message.speaker.actor;
-            const actor = game.actors.get(actorId);
-            console.log('Avant | Found actor by ID:', actorId, '→', actor?.name);
-            console.log('Avant | Actor system data:', actor?.system);
-            console.log('Avant | === _getActorFromMessage END (found) ===');
-            return actor;
-        }
-        
-        console.log('Avant | ERROR: No actor ID found in message speaker');
-        console.log('Avant | === _getActorFromMessage END (not found) ===');
-        return null;
-    }
-    
-    /**
-     * Check if roll is eligible for reroll (2d10 roll)
-     * @param {Roll} roll - The roll to check
-     * @returns {boolean} True if eligible
-     */
-    static _isEligibleRoll(roll) {
-        console.log('Avant | === _isEligibleRoll START ===');
-        console.log('Avant | Roll object:', roll);
-        console.log('Avant | Roll formula:', roll?.formula);
-        console.log('Avant | Roll terms count:', roll?.terms?.length);
-        console.log('Avant | Roll terms:', roll?.terms);
-        
-        if (!roll || !roll.terms || !Array.isArray(roll.terms)) {
-            console.log('Avant | ERROR: Invalid roll object or no terms');
-            console.log('Avant | === _isEligibleRoll END (invalid) ===');
-            return false;
-        }
-        
-        // Check if roll contains exactly 2d10
-        let d10Count = 0;
-        
-        for (let i = 0; i < roll.terms.length; i++) {
-            const term = roll.terms[i];
-            console.log(`Avant | Checking term ${i}:`, term);
-            console.log(`Avant | Term ${i} type:`, term?.constructor?.name);
-            console.log(`Avant | Term ${i} faces:`, term?.faces);
-            console.log(`Avant | Term ${i} number:`, term?.number);
-            
-            if (term instanceof foundry.dice.terms.Die && term.faces === 10) {
-                d10Count += term.number;
-                console.log(`Avant | Term ${i} is d10 with ${term.number} dice, total count now:`, d10Count);
-            } else {
-                console.log(`Avant | Term ${i} is not a d10 die`);
-            }
-        }
-        
-        const eligible = d10Count === 2;
-        console.log('Avant | Final analysis: Roll has', d10Count, 'd10 dice');
-        console.log('Avant | Eligible for reroll:', eligible);
-        console.log('Avant | === _isEligibleRoll END ===');
-        return eligible;
-    }
-}
-
-// CLEAN ACTOR VALIDATION APPROACH
-// Simple hook to ensure type is always set for actor creation
-Hooks.on('preCreateActor', (document, data, options, userId) => {
-    console.log('Avant | preCreateActor hook fired');
-    console.log('Avant | Actor data:', JSON.stringify(data));
-    
-    // Ensure type is set to a valid value
-    if (!data.type || typeof data.type !== 'string' || data.type.trim() === '') {
-        console.log('Avant | Setting default actor type to character');
-        data.type = 'character';
-    }
-    
-    // Validate type is supported
-    const supportedTypes = ['character', 'npc', 'vehicle'];
-    if (!supportedTypes.includes(data.type)) {
-        console.log(`Avant | Invalid actor type '${data.type}', defaulting to character`);
-        data.type = 'character';
-    }
-    
-    // Ensure system data exists
-    if (!data.system) {
-        console.log('Avant | Creating empty system data object');
-        data.system = {};
-    }
-    
-    console.log(`Avant | Actor validation complete - type: ${data.type}`);
-});
-
-Hooks.once('init', async function() {
-    console.log("Avant | Initializing game system");
-    
-    // Register theme manager settings first
-    AvantThemeManager.registerSettings();
-    
-    // Register system settings
-    game.settings.register('avant', 'systemVersion', {
-        name: 'System Version',
-        hint: 'The current version of the Avant system.',
-        scope: 'world',
-        config: false,
-        type: String,
-        default: '0.2.0'
-    });
-
-    // Configure system
-    CONFIG.Actor.documentClass = Actor;
-    CONFIG.Item.documentClass = Item;
-    
-    // Register data models for ALL actor types
-    CONFIG.Actor.dataModels = {
-        character: AvantActorData,
-        npc: AvantActorData,  // Use same model for NPCs
-        vehicle: AvantActorData  // And vehicles if needed
-    };
-    
-    // Register data models for all item types
-    CONFIG.Item.dataModels = {
-        action: AvantActionData,
-        feature: AvantFeatureData,
-        talent: AvantTalentData,
-        augment: AvantAugmentData,
-        weapon: AvantWeaponData,
-        armor: AvantArmorData,
-        gear: AvantGearData
-    };
-    
-    // Register sheet application classes - v12/v13 compatible
-    const ActorsCollection = foundry.documents?.collections?.Actors || Actors;
-    const ItemsCollection = foundry.documents?.collections?.Items || Items;
-    const ActorSheetClass = foundry.appv1?.sheets?.ActorSheet || ActorSheet;
-    const ItemSheetClass = foundry.appv1?.sheets?.ItemSheet || ItemSheet;
-    
-    ActorsCollection.unregisterSheet("core", ActorSheetClass);
-    ActorsCollection.registerSheet("avant", AvantActorSheet, { 
-        makeDefault: true,
-        types: ["character", "npc", "vehicle"]  // Specify supported types
-    });
-    
-    ItemsCollection.unregisterSheet("core", ItemSheetClass);
-    ItemsCollection.registerSheet("avant", AvantItemSheet, { 
-        makeDefault: true,
-        types: ["action", "feature", "talent", "augment", "weapon", "armor", "gear"]
-    });
-    
-    // Preload templates - v12/v13 compatible
-    const loadTemplatesFunc = foundry.applications?.handlebars?.loadTemplates || loadTemplates;
-    loadTemplatesFunc([
-        "systems/avant/templates/actor-sheet.html",
-        "systems/avant/templates/reroll-dialog.html",
-        "systems/avant/templates/item/item-action-sheet.html",
-        "systems/avant/templates/item/item-feature-sheet.html",
-        "systems/avant/templates/item/item-talent-sheet.html",
-        "systems/avant/templates/item/item-augment-sheet.html",
-        "systems/avant/templates/item/item-weapon-sheet.html",
-        "systems/avant/templates/item/item-armor-sheet.html",
-        "systems/avant/templates/item/item-gear-sheet.html"
-    ]);
-    
-    console.log("Avant | System initialization complete");
-});
-
-// System ready
-Hooks.once('ready', async function() {
-    console.log("Avant | System ready");
-    
-    // Initialize theme manager
-    game.avant = game.avant || {};
-    game.avant.themeManager = new AvantThemeManager();
-    
-    // Initialize chat context menu for Fortune Point rerolls with enhanced debugging
-    console.log("Avant | === INITIALIZING CONTEXT MENU SYSTEM ===");
-    console.log("Avant | Using single getChatLogEntryContext approach for v12/v13 compatibility");
-    AvantChatContextMenu.addContextMenuListeners();
-    console.log("Avant | === CONTEXT MENU SYSTEM INITIALIZED ===");
-    
-    ui.notifications.info("Avant game system loaded successfully!");
-});
-
-// Item validation hook
-Hooks.on('preCreateItem', (document, data, options, userId) => {
-    console.log("Avant | preCreateItem called with data:", data);
-    const itemType = data.type;
-    if (itemType === "action" && !data.system?.ability) {
-        document.updateSource({ "system.ability": "might" });
-    }
-    if (itemType === "feature" && !data.system?.category) {
-        document.updateSource({ "system.category": "general" });
-    }
-    if (itemType === "augment" && !data.system?.augmentType) {
-        document.updateSource({ "system.augmentType": "enhancement" });
-    }
-    if (itemType === "weapon") {
-        if (!data.system?.ability) {
-            document.updateSource({ "system.ability": "might" });
-        }
-        if (!data.system?.modifier) {
-            document.updateSource({ "system.modifier": 0 });
-        }
-        if (!data.system?.damageDie) {
-            document.updateSource({ "system.damageDie": "1d6" });
-        }
-        if (!data.system?.threshold) {
-            document.updateSource({ "system.threshold": 11 });
-        }
-    }
-    if (itemType === "armor") {
-        if (!data.system?.ability) {
-            document.updateSource({ "system.ability": "grace" });
-        }
-        if (!data.system?.modifier) {
-            document.updateSource({ "system.modifier": 0 });
-        }
-        if (!data.system?.threshold) {
-            document.updateSource({ "system.threshold": 11 });
-        }
-    }
-    const supportedItemTypes = ["action", "feature", "talent", "augment", "weapon", "armor", "gear"];
-    if (!supportedItemTypes.includes(itemType)) {
-        console.log(`Avant | Unsupported item type '${itemType}', defaulting to 'gear'`);
-        document.updateSource({ type: "gear" });
-    }
-});
-
-// Success logging
-Hooks.on('createActor', (actor, options, userId) => {
-    console.log("Avant | Actor created successfully:", actor.name, "Type:", actor.type);
-});
-
-Hooks.on('createItem', (item, options, userId) => {
-    console.log("Avant | Item created successfully:", item.name, "Type:", item.type);
-});
-
-// Export system API
-window['AVANT'] = {
-    version: '0.2.0',
-    AvantActorData,
-    AvantActionData,
-    AvantFeatureData,
-    AvantTalentData,
-    AvantAugmentData,
-    AvantWeaponData,
-    AvantArmorData,
-    AvantGearData,
-    AvantActorSheet,
-    AvantItemSheet,
-    AvantThemeManager,
-    AvantRerollDialog,
-    AvantChatContextMenu
-};
+console.log("Avant | System loaded successfully - all components imported and configured");
+CompatibilityUtils.log("System load complete - ready for use");
