+++ system.json
@@ -2,7 +2,7 @@
   "id": "avant",
   "title": "Avant",
   "description": "A cutting-edge narrative RPG system focused on experimental storytelling, collaborative worldbuilding, and innovative narrative mechanics",
-  "version": "0.2.0",
+  "version": "0.1.2",
   "compatibility": {
     "minimum": "12.324",
     "verified": "13.341",
