+++ templates/item-sheet.html
@@ -1,0 +1,57 @@
+<form class="{{cssClass}} avant flexcol" autocomplete="off">
+    {{!-- Sheet Header --}}
+    <header class="sheet-header">
+        <img class="profile-img" src="{{item.img}}" data-edit="img" title="{{item.name}}" height="64" width="64"/>
+        <div class="header-fields">
+            <h1 class="charname"><input name="name" type="text" value="{{item.name}}" placeholder="Item Name"/></h1>
+            <div class="resource">
+                <label class="resource-label">Type</label>
+                <select name="type" data-dtype="String">
+                    <option value="action" {{#if (eq item.type "action")}}selected{{/if}}>Action</option>
+                    <option value="feature" {{#if (eq item.type "feature")}}selected{{/if}}>Feature</option>
+                    <option value="talent" {{#if (eq item.type "talent")}}selected{{/if}}>Talent</option>
+                    <option value="augment" {{#if (eq item.type "augment")}}selected{{/if}}>Augment</option>
+                    <option value="weapon" {{#if (eq item.type "weapon")}}selected{{/if}}>Weapon</option>
+                    <option value="armor" {{#if (eq item.type "armor")}}selected{{/if}}>Armor</option>
+                    <option value="gear" {{#if (eq item.type "gear")}}selected{{/if}}>Gear</option>
+                </select>
+            </div>
+        </div>
+    </header>
+
+    {{!-- Sheet Tab Navigation --}}
+    <nav class="sheet-tabs tabs" data-group="primary">
+        <a class="item" data-tab="description">Description</a>
+        <a class="item" data-tab="details">Details</a>
+    </nav>
+
+    {{!-- Sheet Body --}}
+    <section class="sheet-body">
+        {{!-- Description Tab --}}
+        <div class="tab" data-group="primary" data-tab="description">
+            <div class="resource">
+                <label class="resource-label">Description</label>
+                {{editor content=system.description target="system.description" button=true owner=owner editable=editable}}
+            </div>
+        </div>
+
+        {{!-- Details Tab - Dynamic Content Based on Item Type --}}
+        <div class="tab" data-group="primary" data-tab="details">
+            {{#if (eq item.type "action")}}
+                {{> "systems/avant/templates/item/item-action-sheet.html"}}
+            {{else if (eq item.type "feature")}}
+                {{> "systems/avant/templates/item/item-feature-sheet.html"}}
+            {{else if (eq item.type "talent")}}
+                {{> "systems/avant/templates/item/item-talent-sheet.html"}}
+            {{else if (eq item.type "augment")}}
+                {{> "systems/avant/templates/item/item-augment-sheet.html"}}
+            {{else if (eq item.type "weapon")}}
+                {{> "systems/avant/templates/item/item-weapon-sheet.html"}}
+            {{else if (eq item.type "armor")}}
+                {{> "systems/avant/templates/item/item-armor-sheet.html"}}
+            {{else}}
+                {{> "systems/avant/templates/item/item-gear-sheet.html"}}
+            {{/if}}
+        </div>
+    </section>
+</form>
