--- test-context-menu.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Avant Context Menu Test</title>
-</head>
-<body>
-    <h1>🎯 Avant v13 Context Menu Test</h1>
-    
-    <h2>Testing Instructions:</h2>
-    <ol>
-        <li><strong>Open FoundryVTT v13:</strong> <a href="http://localhost:30000" target="_blank">http://localhost:30000</a></li>
-        <li><strong>Create/Open a world</strong> using the Avant system</li>
-        <li><strong>Open character sheet</strong> and ensure character has Fortune Points > 0</li>
-        <li><strong>Make a roll:</strong> Click any ability modifier (e.g., +2 next to Might)</li>
-        <li><strong>Right-click the roll result</strong> in chat</li>
-        <li><strong>Look for:</strong> "Reroll with Fortune Points" option in the context menu</li>
-    </ol>
-    
-    <h2>Expected Results:</h2>
-    <h3>✅ v13 Native Hook (PRIMARY):</h3>
-    <pre>
-Console Output:
-Avant | getChatMessageContextOptions hook fired for message: [message-id]
-Avant | Processing v13 reroll option for message: [message-id]
-Avant | Adding v13 reroll option to context menu
-Avant | v13 reroll option added, total options: [number]
-    </pre>
-    
-    <h3>🔄 v12 Compatibility (FALLBACK):</h3>
-    <pre>
-Console Output:
-Avant | getChatLogEntryContext hook fired (v12 compatibility)
-Avant | Successfully added reroll option
-    </pre>
-    
-    <h3>🛡️ Custom Menu (ULTIMATE FALLBACK):</h3>
-    <pre>
-Console Output:
-Avant | Direct contextmenu event triggered
-Avant | Custom context menu created and displayed
-    </pre>
-    
-    <h2>Context Menu Should Show:</h2>
-    <ul>
-        <li>🎲 <strong>Reroll with Fortune Points</strong> - Native FoundryVTT menu item</li>
-        <li>Click opens familiar reroll dialog</li>
-        <li>Dialog shows individual d10 dice for selection</li>
-        <li>Deducts Fortune Points appropriately</li>
-    </ul>
-    
-    <h2>🐛 Troubleshooting:</h2>
-    <p>If the native context menu doesn't work, check browser console for:</p>
-    <ul>
-        <li>Hook registration messages during system load</li>
-        <li>Hook firing messages when right-clicking</li>
-        <li>Any JavaScript errors</li>
-    </ul>
-    
-    <p><strong>Multiple approaches ensure functionality!</strong> At least one method should work.</p>
-    
-    <hr>
-    <p><em>Generated: v0.3.0 - v13 Native Context Menu Implementation</em></p>
-</body>
-</html>
